Initial move of webapp to a presenter package
