+++ src/webapp/components/app/themes/dhis2-legacy.theme.js
