+++ src/webapp/utils/common.ts
