+++ src/webapp/utils/deprecated.js
