+++ src/webapp/utils/periods.js
