+++ src/webapp/utils/strings.js
